requirements.txt
beautifulsoup4>=4.10.0
click>=8.0.3
Flask>=3.0.0
itsdangerous>=2.1.0
Jinja2>=3.1.6
joblib>=1.2.0
simpledorff>=0.0.2
MarkupSafe>=2.1.0
numpy>=1.21.0
pandas>=1.3.5
python-dateutil>=2.8.2
pytz>=2021.3
PyYAML>=6.0.1
requests>=2.31.0
scikit-learn>=1.0.2
scipy>=1.7.3
six>=1.16.0
soupsieve>=2.3.1
threadpoolctl>=3.0.0
tqdm>=4.62.3
ujson>=5.4.0
Werkzeug>=3.0.6
pydantic>=2.11.9
ollama>=0.6.0
openai>=1.0.0

# Embedding Visualization (optional - for admin dashboard UMAP visualization)
umap-learn>=0.5.0       # UMAP dimensionality reduction

# Extended Format Support (optional - install for PDF, DOCX, etc.)
pdfplumber>=0.10.0      # PDF text extraction with positions
python-docx>=1.0.0      # Word document parsing
mammoth>=1.6.0          # DOCX to HTML conversion
mistune>=3.0.0          # Markdown parsing
pygments>=2.17.0        # Syntax highlighting for code and Markdown
openpyxl>=3.1.0         # Excel file support
